+++ packages/tswebextension/src/lib/mv2/background/cosmetic-api.ts
@@ -1,3 +1,4 @@
+import { nanoid } from 'nanoid';
 import type {
     CosmeticResult,
     CosmeticRule,
@@ -269,7 +270,6 @@ export class CosmeticApi {
             frameId,
             cosmeticResult,
             url,
-            requestId,
             contentType,
             timestamp,
         } = params;
@@ -296,7 +296,10 @@ export class CosmeticApi {
                         data: {
                             script: true,
                             tabId,
-                            eventId: requestId,
+                            // for proper filtering log request info rule displaying
+                            // event id should be unique for each event, not copied from request
+                            // https://github.com/AdguardTeam/AdguardBrowserExtension/issues/2341
+                            eventId: nanoid(),
                             requestUrl: url,
                             frameUrl: url,
                             frameDomain: getDomain(url) as string,