+++ packages/tswebextension/CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- TODO: manually add compare links for version to the end of the file -->
 <!-- e.g. [0.1.2]: https://github.com/AdguardTeam/tsurlfilter/compare/tswebextension-v0.1.1...tswebextension-v0.1.2 -->
 
+## Unreleased
+
+### Fixed
+
+- unique `eventId` for `FilteringEventType.JsInject` events
+
+
 ## [0.1.4] - 2023-04-18
 
 ### Fixed